Přesunuty dávky do nového blueprintu
