--- registry/templates/donor/batch_list.html
@@ -1,50 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-
-<h1>Přehled importovaných dávek</h1>
-
-<table id="batch_list" class="table table-striped table-hovered table-hover">
-    <thead class="thead-dark">
-        <tr>
-            <th>Číslo</th>
-            <th>Importováno</th>
-            <th>Odběrné místo</th>
-            <th>Odstranit</th>
-        </tr>
-    </thead>
-    <tbody>
-        {% for batch in batches %}
-            <tr>
-                <td><a href="{{ url_for('donor.batch_detail', id=batch.id) }}">{{ batch.id}}</a></td>
-                <td>{{ batch.imported_at}}</td>
-                <td>{{ batch.donation_center.title or "Jinde"}}</td>
-                <td>
-                    {% with form=delete_batch_form %}
-                    <form id="deleteBatchForm" action="{{ url_for('donor.delete_batch') }}" method="POST"
-                        class="form-inline" role="form">
-                        <div class="form-group">
-                            {{ form.csrf_token }}
-                            {{ form.batch_id(value=batch.id) }}
-                            <input type="submit" class="btn btn-sm btn-danger" value="🗑">
-                        </div>
-                    </form>
-                    {% endwith %}
-                </td>
-            </tr>
-        {% endfor %}
-    </tbody>
-</table>
-
-{% endblock %}
-
-{% block js %}
-<script type="text/javascript">
-    $(document).ready( function () {
-        $('#batch_list').DataTable({
-            language: {
-                url: '//cdn.datatables.net/plug-ins/1.10.21/i18n/Czech.json'
-            },
-        });
-    } );
-</script>
-{% endblock %}
