+++ registry/templates/nav.html
@@ -10,10 +10,10 @@
       <a class="nav-link" href="{{ url_for('donor.overview') }}">Přehled dárců</a>
     </li>
     <li class="nav-item">
-      <a class="nav-link" href="{{ url_for('donor.batch_list') }}">Přehled importů</a>
+      <a class="nav-link" href="{{ url_for('batch.batch_list') }}">Přehled importů</a>
     </li>
     <li class="nav-item">
-      <a class="nav-link" href="{{ url_for('donor.import_data') }}">Import dat</a>
+      <a class="nav-link" href="{{ url_for('batch.import_data') }}">Import dat</a>
     </li>
 
     <li class="nav-item dropdown">
@@ -22,7 +22,7 @@
       </a>
       <div class="dropdown-menu">
         {% for medal in all_medals %}
-          <a class="dropdown-item" href="{{ url_for('donor.award_prep', medal_slug=medal.slug)}}">{{ medal.title }}</a>
+        <a class="dropdown-item" href="{{ url_for('donor.award_prep', medal_slug=medal.slug)}}">{{ medal.title }}</a>
         {% endfor %}
       </div>
     </li>
@@ -31,23 +31,24 @@
 
   <ul class="navbar-nav ml-auto">
     {% if current_user and current_user.is_authenticated %}
-      <li class="nav-item">
-        <span class="navbar-text">
-          Přihlášen jako {{ current_user.email }}
-        </span>
-        <a class="btn btn-info" role="button" href="{{ url_for('public.logout') }}">Odhlásit se</a>
-      </li>
+    <li class="nav-item">
+      <span class="navbar-text">
+        Přihlášen jako {{ current_user.email }}
+      </span>
+      <a class="btn btn-info" role="button" href="{{ url_for('public.logout') }}">Odhlásit se</a>
+    </li>
     {% else %}
-      {% with form=form %}
-      <form class="form-inline" id="loginForm" method="POST" action="/" role="login">
-          {{ form.csrf_token }}
-          <div>
-            {{ form.email(placeholder="E-mail", class_="form-control mr-sm-2 my-auto") }} {{ form.password(placeholder="Heslo",
-            class_="form-control mr-sm-2 my-auto") }}
-            <button class="btn btn-success" type="submit">Přihlásit</button>
-          </div>
-      </form>
-      {% endwith %}
+    {% with form=form %}
+    <form class="form-inline" id="loginForm" method="POST" action="/" role="login">
+      {{ form.csrf_token }}
+      <div>
+        {{ form.email(placeholder="E-mail", class_="form-control mr-sm-2 my-auto") }} {{
+        form.password(placeholder="Heslo",
+        class_="form-control mr-sm-2 my-auto") }}
+        <button class="btn btn-success" type="submit">Přihlásit</button>
+      </div>
+    </form>
+    {% endwith %}
     {% endif %}
   </ul>
 </nav>
